pyproject.toml
[project]
name = "FNN-news-recommender"
version = "0.1.0"
description = "A Fuzzy Neural Network implementation for news recommendation. Code of the research paper `Modeling Behavioral Patterns in News Recommendations Using Fuzzy Neural Networks`"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "datasets>=4.0.0",
    "ebrec",
    "numpy<2",
    "recommenders>=1.2.1",
    "scikit-learn==1.4.0",
    "scikit-posthocs>=0.11.4",
    "scikit-surprise>=1.1.4",
    "seaborn>=0.13.2",
    "simple-parsing>=0.1.7",
    "torch==2.2.2",
    "torch-geometric>=2.6.1",
    "torchaudio==2.2.2",
    "torchvision==0.17.2",
]
classifiers = ["Private :: Do Not Upload"]

[tool.uv.sources]
ebrec = { git = "https://github.com/ebanalyse/ebnerd-benchmark.git" }
